feat(common): Add ExternalId class to represent an external id
